apps/code-diff-review/requirements.txt
cururo
requests
PyGithub